Move apply-effect handler to `ChatLogPF2e`, fix aura-traits restriction
