+++ packs/feat-effects/effect-kinetic-aura.json
@@ -87,7 +87,6 @@
                 "radius": "@item.flags.pf2e.rulesSelections.auraRadius",
                 "slug": "kinetic-aura",
                 "traits": [
-                    "kineticist",
                     "primal"
                 ]
             },
