+++ static/templates/chat/action/collapsed.hbs
@@ -19,7 +19,7 @@
             </div>
         {{/if}}
         {{#if selfEffect}}
-            <button type="button" data-action="apply-effect" data-targets="{{actor.uuid}}" data-visibility="owner">
+            <button type="button" data-action="applyEffect" data-targets="{{actor.uuid}}" data-visibility="owner">
                 {{localize "PF2E.Action.ApplyEffect"}}
             </button>
         {{/if}}
